Create thumbnails in Profile.save (#2903)
* Create thumbnails in Profile.save

* Use update_image

* Move tests into models_test.py

* Lint
